add verify for recommend level limit

+++ Rebate-Mechanism/rebate-pom/rebate-interface/src/main/java/org/rebate/controller/EndUserController.java
@@ -12,6 +12,7 @@ import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.commons.lang.BooleanUtils;
 import org.rebate.aspect.UserValidCheck;
 import org.rebate.beans.CommonAttributes;
 import org.rebate.beans.Message;
@@ -546,7 +547,13 @@ public class EndUserController extends MobileBaseController {
           response.setCode(CommonAttributes.FAIL_REG);
           response.setDesc(Message.error("rebate.sms.token.error").getContent());
           return response;
-        } else {
+        } else {// 短信验证码匹配成功
+          if (BooleanUtils.isTrue(endUserService.isRecommendLimited(recommenderMobile))) {// 推荐层级关系超过配置的最大层级
+            response.setCode(CommonAttributes.FAIL_REG);
+            response.setDesc(Message.error("rebate.recommend.level.limit", recommenderMobile)
+                .getContent());
+            return response;
+          }
           endUserService.deleteSmsCode(cellPhoneNum);
         }
       } else {
